fix: resolve CI formatting and type checks
