+++ src/tempo_dag/examples/__init__.py
@@ -1,0 +1,15 @@
+from .temporal_demo import (
+    OUTPUT_DIR,
+    RollingMeanConvDemoModel,
+    TemporalDemoReport,
+    TemporalStepMetric,
+    run_demo,
+)
+
+__all__ = [
+    "OUTPUT_DIR",
+    "RollingMeanConvDemoModel",
+    "TemporalDemoReport",
+    "TemporalStepMetric",
+    "run_demo",
+]
