+++ tests/unit/test_temporal_hls.py
@@ -10,10 +10,14 @@ from tempo_dag.verification.golden_trace import load_golden_trace
 
 
 def test_render_temporal_process_hls_includes_buffers_and_step_function() -> None:
-    process = TemporalONNXParser().parse_model(
-        build_demo_temporal_onnx_model(),
-        process_id="demo_process",
-    ).process
+    process = (
+        TemporalONNXParser()
+        .parse_model(
+            build_demo_temporal_onnx_model(),
+            process_id="demo_process",
+        )
+        .process
+    )
 
     rendered = render_temporal_process_hls(process)
 
@@ -23,10 +27,14 @@ def test_render_temporal_process_hls_includes_buffers_and_step_function() -> Non
 
 
 def test_render_temporal_artifact_includes_testbench_trace_comments() -> None:
-    process = TemporalONNXParser().parse_model(
-        build_demo_temporal_onnx_model(),
-        process_id="demo_process",
-    ).process
+    process = (
+        TemporalONNXParser()
+        .parse_model(
+            build_demo_temporal_onnx_model(),
+            process_id="demo_process",
+        )
+        .process
+    )
     trace = load_golden_trace("tests/verification/golden_traces/rolling_mean.json")
 
     artifact = render_temporal_artifact_from_trace(process, trace)
